PEEK-614: bug: Change worker plpython method to use SD not GD
Using GD caused problems with the SQLAlchemy mapper when other
statements had been run in the python client prior to the
worker task.

EG, the peek-plugin-diagrams DispBase LiveDbDispLink mapping which
must be explicitly loaded with loadStorageTuples() when accessing
DispBase - which conflicted with the Queue Compiler calls as it 
didn't need those imports.